requirements.txt
python-dotenv
requests
pinecone
langchain
langchain-pinecone==0.2.0
langchain-openai
langgraph
langchain-community
pyTelegramBotAPI
langgraph-checkpoint-sqlite